VideoEdgeSolution/modules/ImageAnalyzer/requirements.txt
azure-iot-device~=2.0.0
azure-iothub-device-client
azure-storage-blob
requests